+++ app/src/main/java/uniandes/edu/co/app/repositorio/HabitacionRepo.java
@@ -32,14 +32,14 @@ public interface HabitacionRepo extends JpaRepository<Habitacion, Integer> {
         @Query(value = "UPDATE habitaciones SET capacidad = :capacidad, tipo = :tipo, dotacion = :dotacion, precioNoche = :precioNoche, consumoExtra = :consumoExtra WHERE id = :id", nativeQuery = true)
         void actualizarHabitacion(@Param("id") long id, @Param("capacidad") Integer capacidad,
                         @Param("tipo") String tipo,
-                        @Param("dotacion") String dotacion, @Param("precioNoche") Double precioNoche,
-                        @Param("consumoExtra") Double consumoExtra);
+                        @Param("dotacion") String dotacion, @Param("precioNoche") Integer precioNoche,
+                        @Param("consumoExtra") Integer consumoExtra);
 
         @Modifying
         @Transactional
         @Query(value = "INSERT INTO habitaciones (id, capacidad, tipo, dotacion, precioNoche, consumoExtra) VALUES ( hotelandes_sequence.nextval , :capacidad, :tipo, :dotacion, :precioNoche, :consumoExtra)", nativeQuery = true)
         void insertarHabitacion(@Param("capacidad") Integer capacidad, @Param("tipo") String tipo,
-                        @Param("dotacion") String dotacion, @Param("precioNoche") Double precioNoche,
-                        @Param("consumoExtra") Double consumoExtra);
+                        @Param("dotacion") String dotacion, @Param("precioNoche") Integer precioNoche,
+                        @Param("consumoExtra") Integer consumoExtra);
 
 }
